Move the update link timestamp route from items into links
